Доработано разделение на модули
