+++ js/main.js
@@ -1,4 +1,4 @@
-import { generatePhotos } from './data.js';
+import { generatePhotos } from './photosDataGeneration.js';
 
 // eslint-disable-next-line no-console
 console.log(generatePhotos());
